+++ ui-design-lib/src/main/java/co/electriccoin/zcash/ui/design/component/ZashiQr.kt
@@ -2,14 +2,18 @@ package co.electriccoin.zcash.ui.design.component
 
 import androidx.compose.foundation.BorderStroke
 import androidx.compose.foundation.Image
+import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material3.Surface
 import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.ImageBitmap
+import androidx.compose.ui.graphics.painter.Painter
 import androidx.compose.ui.platform.LocalConfiguration
 import androidx.compose.ui.platform.LocalDensity
 import androidx.compose.ui.unit.Dp
@@ -26,7 +30,10 @@ fun ZashiQr(
     qrData: String,
     modifier: Modifier = Modifier,
     qrSize: Dp = ZashiQrDefaults.width,
-    colors: QrCodeColors = QrCodeDefaults.colors()
+    colors: QrCodeColors = QrCodeDefaults.colors(),
+    contentDescription: String? = null,
+    onQrCodeClick: () -> Unit = {},
+    centerImage: Painter? = null
 ) {
     val qrSizePx = with(LocalDensity.current) { qrSize.roundToPx() }
     val bitmap = getQrCode(qrData, qrSizePx, colors)
@@ -35,15 +42,26 @@ fun ZashiQr(
         modifier = modifier,
         shape = RoundedCornerShape(ZashiDimensions.Radius.radius4xl),
         border = BorderStroke(width = 1.dp, color = ZashiColors.Surfaces.strokePrimary),
-        color = ZashiColors.Surfaces.bgPrimary
+        color = ZashiColors.Surfaces.bgPrimary,
     ) {
         Box(
             modifier = Modifier.padding(all = 12.dp)
         ) {
             Image(
                 bitmap = bitmap,
-                contentDescription = null,
+                contentDescription = contentDescription,
+                Modifier.clickable { onQrCodeClick() }
             )
+            if (centerImage != null) {
+                Image(
+                    modifier =
+                        Modifier
+                            .size(64.dp)
+                            .align(Alignment.Center),
+                    painter = centerImage,
+                    contentDescription = contentDescription,
+                )
+            }
         }
     }
 }