+++ ui-lib/src/main/java/co/electriccoin/zcash/ui/Navigation.kt
@@ -541,20 +541,7 @@ fun NavHostController.popBackStackJustOnce(currentRouteToBePopped: String) {
 object NavigationArguments {
     const val SEND_SCAN_RECIPIENT_ADDRESS = "send_scan_recipient_address"
     const val SEND_SCAN_ZIP_321_URI = "send_scan_zip_321_uri"
-
-    const val SEND_CONFIRM_RECIPIENT_ADDRESS = "send_confirm_recipient_address"
-    const val SEND_CONFIRM_AMOUNT = "send_confirm_amount"
-    const val SEND_CONFIRM_MEMO = "send_confirm_memo"
-    const val SEND_CONFIRM_PROPOSAL = "send_confirm_proposal"
-    const val SEND_CONFIRM_INITIAL_STAGE = "send_confirm_initial_stage"
-
     const val MULTIPLE_SUBMISSION_CLEAR_FORM = "multiple_submission_clear_form"
-
-    const val PAYMENT_REQUEST_ADDRESS = "payment_request_address"
-    const val PAYMENT_REQUEST_AMOUNT = "payment_request_amount"
-    const val PAYMENT_REQUEST_MEMO = "payment_request_memo"
-    const val PAYMENT_REQUEST_PROPOSAL = "payment_request_proposal"
-    const val PAYMENT_REQUEST_URI = "payment_request_uri"
 }
 
 object NavigationTargets {