+++ ui-lib/src/main/java/co/electriccoin/zcash/ui/screen/qrcode/model/QrCodeState.kt
@@ -3,7 +3,7 @@ package co.electriccoin.zcash.ui.screen.qrcode.model
 import androidx.compose.ui.graphics.ImageBitmap
 import cash.z.ecc.android.sdk.model.WalletAddress
 
-internal sealed class QrCodeState {
+sealed class QrCodeState {
     data object Loading : QrCodeState()
 
     data class Prepared(
@@ -11,6 +11,7 @@ internal sealed class QrCodeState {
         val walletAddress: WalletAddress,
         val onAddressCopy: (String) -> Unit,
         val onQrCodeShare: (ImageBitmap) -> Unit,
+        val onQrCodeClick: () -> Unit,
         val onBack: () -> Unit,
     ) : QrCodeState()
 }